Code/requirements_dashboard.txt
# ASL Alphabet Recognition Dashboard Requirements
# Install with: pip install -r requirements_dashboard.txt

# Core Framework
streamlit>=1.28.0

# Deep Learning
torch>=1.12.0
torchvision>=0.13.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Utilities
python-dateutil>=2.8.2
